Remove loop over groups in the custom regularizer (for semi-sparse group lasso)
In SSGL_WeightRegularizer a loop over the groups was used to compute
the loss function (regularization term on groups).

Since the number of features can be very large (>1000) in the context
of the Random Shapelet -and thus the number of groups- the training was
very long. The loop is replaced by tensorflow.segment_sum function and
the prepare_groups function is now useless.

The training time is now in the seconds instead of 10+ minutes.

In discussion with Alexis Gatignol (alexis.gatignol@axa.com)